requirements.txt
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langsmith>=0.0.50
groq>=0.4.0
twilio>=8.0.0
python-dotenv>=1.0.0
flask>=3.0.0
pandas>=2.0.0
